backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
requests==2.32.3
gunicorn==21.2.0
whitenoise==6.8.2
